more async refactoring (#219)
* break up worker.go and worker_test.go into multiple files

* simplify heart

* simplify cleaner

* move heart from worker to engine

* remove receive/execute/watch from worker to engine

* let task receivers return right after fatal error

* let task executor return right after fatal error

* very minor comment cleanup

* fix lint errors by removing unused error types
