+++ pkg/async/redis/engine_test.go
@@ -5,34 +5,34 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Azure/open-service-broker-azure/pkg/async/redis/fake"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestEngineRunBlocksUntilCleanerStops(t *testing.T) {
-	e := NewEngine(redisClient).(*engine)
+func TestNewEnginesHaveUniqueWorkerIDs(t *testing.T) {
+	// Create two engines
+	e1 := NewEngine(redisClient).(*engine)
+	e2 := NewEngine(redisClient).(*engine)
 
-	// Create a fake cleaner that will just return an error when it runs
-	c := fake.NewCleaner()
-	c.RunBehavior = func(context.Context) error {
+	// Assert that their workerIDs are at least different from one another
+	assert.NotEqual(t, e1.workerID, e2.workerID)
+}
+
+func TestRunBlocksUntilCleanReturnsError(t *testing.T) {
+	e := getTestEngine()
+
+	// Override the engine's clean function so it just returns an error
+	e.clean = func(context.Context, string, string, string) error {
 		return errSome
 	}
 
-	// Make the engine use the fake cleaner
-	e.cleaner = c
-
-	// Create a fake worker that will just communicate when the context it was
-	// passed has been canceled
+	// Override the engine's runHeart function so it just communicates when the
+	// context it was passed has been canceled
 	contextCanceledCh := make(chan struct{})
-	w := fake.NewWorker()
-	w.RunBehavior = func(ctx context.Context) error {
+	e.runHeart = func(ctx context.Context) error {
 		<-ctx.Done()
 		close(contextCanceledCh)
 		return ctx.Err()
 	}
-
-	// Make the engine use the fake worker
-	e.worker = w
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -45,7 +45,7 @@ func TestEngineRunBlocksUntilCleanerStops(t *testing.T) {
 	}()
 
 	// Assert that the error returned from the Run function wraps the error that
-	// the fake heart generated
+	// the overridden clean function returned
 	select {
 	case err := <-errCh:
 		assert.Equal(t, &errCleanerStopped{err: errSome}, err)
@@ -63,30 +63,22 @@ func TestEngineRunBlocksUntilCleanerStops(t *testing.T) {
 	}
 }
 
-func TestEngineRunBlocksUntilWorkerStops(t *testing.T) {
-	e := NewEngine(redisClient).(*engine)
+func TestRunBlocksUntilRunHeartReturnsError(t *testing.T) {
+	e := getTestEngine()
 
-	// Create a fake cleaner that will just communicate when the context it was
-	// passed has been canceled
+	// Override the engine's clean function so it just communicates when the
+	// context it was passed has been canceled
 	contextCanceledCh := make(chan struct{})
-	c := fake.NewCleaner()
-	c.RunBehavior = func(ctx context.Context) error {
+	e.clean = func(ctx context.Context, _ string, _ string, _ string) error {
 		<-ctx.Done()
 		close(contextCanceledCh)
 		return ctx.Err()
 	}
 
-	// Make the engine use the fake cleaner
-	e.cleaner = c
-
-	// Create a fake worker that will just return an error when it runs
-	w := fake.NewWorker()
-	w.RunBehavior = func(context.Context) error {
+	// Override the engine's runHeart function so it just returns an error
+	e.runHeart = func(context.Context) error {
 		return errSome
 	}
-
-	// Make the engine use the fake worker
-	e.worker = w
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -99,16 +91,16 @@ func TestEngineRunBlocksUntilWorkerStops(t *testing.T) {
 	}()
 
 	// Assert that the error returned from the Run function wraps the error that
-	// the fake heart generated
+	// the overridden runHeart function returned
 	select {
 	case err := <-errCh:
-		assert.Equal(t, &errWorkerStopped{workerID: w.GetID(), err: errSome}, err)
+		assert.Equal(t, &errHeartStopped{workerID: e.workerID, err: errSome}, err)
 	case <-time.After(time.Second):
 		assert.Fail(t, "an error should have been received, but wasn't")
 	}
 
 	// Assert that the context got canceled. It's helpful to know that when the
-	// worker stops, the rest of the engine components are also signaled to shut
+	// cleaner stops, the rest of the engine components are also signaled to shut
 	// down.
 	select {
 	case <-contextCanceledCh:
@@ -117,30 +109,276 @@ func TestEngineRunBlocksUntilWorkerStops(t *testing.T) {
 	}
 }
 
-func TestEngineRunRespondsToContextCanceled(t *testing.T) {
-	e := NewEngine(redisClient).(*engine)
+func TestRunBlocksUntilReceivePendingTasksSendsError(t *testing.T) {
+	e := getTestEngine()
 
-	// Create a fake cleaner that will just run until the context it was passed
-	// is canceled
-	c := fake.NewCleaner()
-	c.RunBehavior = func(ctx context.Context) error {
+	// Override the engine's clean function so it just communicates when the
+	// context it was passed has been canceled
+	contextCanceledCh := make(chan struct{})
+	e.clean = func(ctx context.Context, _ string, _ string, _ string) error {
 		<-ctx.Done()
+		close(contextCanceledCh)
 		return ctx.Err()
 	}
 
-	// Make the engine use the fake cleaner
-	e.cleaner = c
+	// Override the engine's receivePendingTasks function so it just sends an
+	// error
+	e.receivePendingTasks = func(
+		ctx context.Context,
+		_ string,
+		_ string,
+		_ chan []byte,
+		errCh chan error,
+	) {
+		select {
+		case errCh <- errSome:
+		case <-ctx.Done():
+		}
+	}
 
-	// Create a fake worker that will just run until the context it was passed
-	// is canceled
-	w := fake.NewWorker()
-	w.RunBehavior = func(ctx context.Context) error {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// Call Run in a goroutine. If it never unblocks, as we hope it does, we don't
+	// want the test to stall.
+	errCh := make(chan error)
+	go func() {
+		errCh <- e.Run(ctx)
+	}()
+
+	// Assert that the error returned from the Run function wraps the error that
+	// the overridden receivePendingTasks function sent
+	select {
+	case err := <-errCh:
+		assert.Equal(
+			t,
+			&errReceiverStopped{
+				workerID:  e.workerID,
+				queueName: pendingTaskQueueName,
+				err:       errSome,
+			},
+			err,
+		)
+	case <-time.After(time.Second):
+		assert.Fail(t, "an error should have been received, but wasn't")
+	}
+
+	// Assert that the context got canceled. It's helpful to know that when the
+	// pending task receiver stops, the rest of the worker components are also
+	// signaled to shut down.
+	select {
+	case <-contextCanceledCh:
+	case <-time.After(time.Second):
+		assert.Fail(t, "context should have been canceled, but it was not")
+	}
+}
+
+func TestRunBlocksUntilExecuteTasksSendsError(t *testing.T) {
+	e := getTestEngine()
+
+	// Override the engine's clean function so it just communicates when the
+	// context it was passed has been canceled
+	contextCanceledCh := make(chan struct{})
+	e.clean = func(ctx context.Context, _ string, _ string, _ string) error {
 		<-ctx.Done()
+		close(contextCanceledCh)
 		return ctx.Err()
 	}
 
-	// Make the engine use the fake worker
-	e.worker = w
+	// Override the engine's executeTasks function so it just sends an error
+	e.executeTasks = func(
+		ctx context.Context,
+		_ chan []byte,
+		_ string,
+		_ string,
+		errCh chan error,
+	) {
+		select {
+		case errCh <- errSome:
+		case <-ctx.Done():
+		}
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// Call Run in a goroutine. If it never unblocks, as we hope it does, we don't
+	// want the test to stall.
+	errCh := make(chan error)
+	go func() {
+		errCh <- e.Run(ctx)
+	}()
+
+	// Assert that the error returned from the Run function wraps the error that
+	// the overridden executeTasks function sent
+	select {
+	case err := <-errCh:
+		assert.Equal(
+			t,
+			&errTaskExecutorStopped{
+				workerID: e.workerID,
+				err:      errSome,
+			},
+			err,
+		)
+	case <-time.After(time.Second):
+		assert.Fail(t, "an error should have been received, but wasn't")
+	}
+
+	// Assert that the context got canceled. It's helpful to know that when the
+	// task executor stops, the rest of the worker components are also signaled to
+	// shut down.
+	select {
+	case <-contextCanceledCh:
+	case <-time.After(time.Second):
+		assert.Fail(t, "context should have been canceled, but it was not")
+	}
+}
+
+func TestRunBlocksUntilReceiveDeferredTasksSendError(t *testing.T) {
+	e := getTestEngine()
+
+	// Override the engine's clean function so it just communicates when the
+	// context it was passed has been canceled
+	contextCanceledCh := make(chan struct{})
+	e.clean = func(ctx context.Context, _ string, _ string, _ string) error {
+		<-ctx.Done()
+		close(contextCanceledCh)
+		return ctx.Err()
+	}
+
+	// Override the engine's receiveDeferredTasks function so it just sends an
+	// error
+	e.receiveDeferredTasks = func(
+		ctx context.Context,
+		_ string,
+		_ string,
+		_ chan []byte,
+		errCh chan error,
+	) {
+		select {
+		case errCh <- errSome:
+		case <-ctx.Done():
+		}
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// Call Run in a goroutine. If it never unblocks, as we hope it does, we don't
+	// want the test to stall.
+	errCh := make(chan error)
+	go func() {
+		errCh <- e.Run(ctx)
+	}()
+
+	// Assert that the error returned from the Run function wraps the error that
+	// the overridden receiveDeferredTasks function sent
+	select {
+	case err := <-errCh:
+		assert.Equal(
+			t,
+			&errReceiverStopped{
+				workerID:  e.workerID,
+				queueName: deferredTaskQueueName,
+				err:       errSome,
+			},
+			err,
+		)
+	case <-time.After(time.Second):
+		assert.Fail(t, "an error should have been received, but wasn't")
+	}
+
+	// Assert that the context got canceled. It's helpful to know that when the
+	// deferred task receiver stops, the rest of the worker components are also
+	// signaled to shut down.
+	select {
+	case <-contextCanceledCh:
+	case <-time.After(time.Second):
+		assert.Fail(t, "context should have been canceled, but it was not")
+	}
+}
+
+func TestRunBlocksUntilWatchDeferredTaskSendsError(t *testing.T) {
+	e := getTestEngine()
+
+	// Override the engine's receiveDeferredTasks function so it just sends a
+	// result (to trigger the logic that spawns a new deferred task watcher) and
+	// then blocks until the context it was passed is canceled, which it will
+	// communicate
+	contextCanceledCh := make(chan struct{})
+	e.receiveDeferredTasks = func(
+		ctx context.Context,
+		_ string,
+		_ string,
+		retCh chan []byte,
+		_ chan error,
+	) {
+		select {
+		case retCh <- []byte{}: // A dummy value is fine
+		case <-ctx.Done():
+		}
+		close(contextCanceledCh)
+		<-ctx.Done()
+	}
+
+	// Override the engine's watchDeferredTask function so it just sends an error
+	e.watchDeferredTask = func(
+		ctx context.Context,
+		_ []byte,
+		_ string,
+		errCh chan error,
+	) {
+		select {
+		case errCh <- errSome:
+		case <-ctx.Done():
+		}
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// Call Run in a goroutine. If it never unblocks, as we hope it does, we don't
+	// want the test to stall.
+	errCh := make(chan error)
+	go func() {
+		errCh <- e.Run(ctx)
+	}()
+
+	// Assert that the error returned from the Run function wraps the error that
+	// the overridden watchDeferredTask function sent
+	select {
+	case err := <-errCh:
+		assert.Equal(
+			t,
+			&errDeferredTaskWatcherStopped{workerID: e.workerID, err: errSome},
+			err,
+		)
+	case <-time.After(time.Second):
+		assert.Fail(t, "an error should have been received, but wasn't")
+	}
+
+	// Assert that the context got canceled. It's helpful to know that when the
+	// a deferred task watcher errors, the rest of the worker components are also
+	// signaled to shut down.
+	select {
+	case <-contextCanceledCh:
+	case <-time.After(time.Second):
+		assert.Fail(t, "context should have been canceled, but it was not")
+	}
+}
+
+func TestRunRespondsToCanceledContext(t *testing.T) {
+	e := getTestEngine()
+
+	// Override the engine's clean function so it just communicates when the
+	// context it was passed has been canceled
+	contextCanceledCh := make(chan struct{})
+	e.clean = func(ctx context.Context, _ string, _ string, _ string) error {
+		<-ctx.Done()
+		close(contextCanceledCh)
+		return ctx.Err()
+	}
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
@@ -165,4 +403,62 @@ func TestEngineRunRespondsToContextCanceled(t *testing.T) {
 			"a context canceled error should have been returned, but wasn't",
 		)
 	}
+}
+
+// getTestEngine returns a pointer to an engine that has all its long-running
+// concurrent functions pre-overridden to simply block until the context they
+// are passed is canceled. Individual test cases can selectively revert or
+// amend these overrides to test specific scenarios.
+func getTestEngine() *engine {
+	e := NewEngine(redisClient).(*engine)
+	// Cleaner loop
+	e.clean = func(ctx context.Context, _ string, _ string, _ string) error {
+		<-ctx.Done()
+		return ctx.Err()
+	}
+	// Heartbeat loop
+	e.runHeart = func(ctx context.Context) error {
+		<-ctx.Done()
+		return ctx.Err()
+	}
+	// Pending tasks receiver
+	e.receivePendingTasks = func(
+		ctx context.Context,
+		_ string,
+		_ string,
+		_ chan []byte,
+		_ chan error,
+	) {
+		<-ctx.Done()
+	}
+	// Deferred tasks receiver
+	e.receiveDeferredTasks = func(
+		ctx context.Context,
+		_ string,
+		_ string,
+		_ chan []byte,
+		_ chan error,
+	) {
+		<-ctx.Done()
+	}
+	// Tasks executor
+	e.executeTasks = func(
+		ctx context.Context,
+		_ chan []byte,
+		_ string,
+		_ string,
+		_ chan error,
+	) {
+		<-ctx.Done()
+	}
+	// Deferred task watcher
+	e.watchDeferredTask = func(
+		ctx context.Context,
+		_ []byte,
+		_ string,
+		errCh chan error,
+	) {
+		<-ctx.Done()
+	}
+	return e
 }
