+++ pkg/async/redis/errors.go
@@ -2,44 +2,12 @@ package redis
 
 import "fmt"
 
-type errCleaning struct {
-	workerID string
-	err      error
-}
-
-func (e *errCleaning) Error() string {
-	if e.workerID == "" {
-		return fmt.Sprintf(
-			`error cleaning up after dead workers: %s`,
-			e.err,
-		)
-	}
-	return fmt.Sprintf(
-		`error cleaning up after dead worker "%s": %s`,
-		e.workerID,
-		e.err,
-	)
-}
-
 type errCleanerStopped struct {
 	err error
 }
 
 func (e *errCleanerStopped) Error() string {
 	baseMsg := "cleaner stopped"
-	if e.err == nil {
-		return baseMsg
-	}
-	return fmt.Sprintf("%s: %s", baseMsg, e.err)
-}
-
-type errWorkerStopped struct {
-	workerID string
-	err      error
-}
-
-func (e *errWorkerStopped) Error() string {
-	baseMsg := fmt.Sprintf(`worker "%s" stopped`, e.workerID)
 	if e.err == nil {
 		return baseMsg
 	}
@@ -112,17 +80,4 @@ type errDuplicateJob struct {
 
 func (e *errDuplicateJob) Error() string {
 	return fmt.Sprintf(`duplicate job name "%s"`, e.name)
-}
-
-type errHeartbeat struct {
-	workerID string
-	err      error
-}
-
-func (e *errHeartbeat) Error() string {
-	return fmt.Sprintf(
-		`error sending heartbeat for worker "%s": %s`,
-		e.workerID,
-		e.err,
-	)
 }
