+++ pkg/async/redis/task_watcher.go
@@ -1,0 +1,87 @@
+package redis
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	log "github.com/Sirupsen/logrus"
+)
+
+// watchDeferredTaskFn defines functions used to watch a deferred task
+type watchDeferredTaskFn func(
+	ctx context.Context,
+	taskJSON []byte,
+	pendingTaskQueueName string,
+	errCh chan error,
+)
+
+func (e *engine) defaultWatchDeferredTask(
+	ctx context.Context,
+	taskJSON []byte,
+	pendingTaskQueueName string,
+	errCh chan error,
+) {
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+	task, err := e.getTaskFromJSON(taskJSON, getWatchedTaskQueueName(e.workerID))
+	if err != nil {
+		select {
+		case errCh <- err:
+		case <-ctx.Done():
+		}
+		return
+	}
+	if task == nil {
+		return
+	}
+	executeTime := task.GetExecuteTime()
+	if executeTime == nil {
+		err := e.redisClient.LRem(
+			getWatchedTaskQueueName(e.workerID),
+			-1,
+			taskJSON,
+		).Err()
+		if err != nil {
+			select {
+			case errCh <- fmt.Errorf(
+				`error removing task "%s" with no executeTime from queue "%s": %s`,
+				task.GetID(),
+				getWatchedTaskQueueName(e.workerID),
+				err,
+			):
+			case <-ctx.Done():
+			}
+			return
+		}
+		log.WithFields(log.Fields{
+			"task":  task.GetID(),
+			"queue": getWatchedTaskQueueName(e.workerID),
+		}).Error("deferred task had no executeTime and was removed from the queue")
+		return
+	}
+	// Note if the duration passed to the timer is 0 or negative, it should go
+	// off immediately
+	timer := time.NewTimer(time.Until(*executeTime))
+	defer timer.Stop()
+	select {
+	case <-timer.C:
+		// Move the task to the pending queue
+		pipeline := e.redisClient.TxPipeline()
+		pipeline.LPush(pendingTaskQueueName, taskJSON)
+		pipeline.LRem(getWatchedTaskQueueName(e.workerID), -1, taskJSON)
+		_, err := pipeline.Exec()
+		if err != nil {
+			select {
+			case errCh <- fmt.Errorf(
+				`error moving deferred task "%s" to queue "%s": %s`,
+				task.GetID(),
+				pendingTaskQueueName,
+				err,
+			):
+			case <-ctx.Done():
+			}
+		}
+	case <-ctx.Done():
+	}
+}
